requirements.txt
# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
aiofiles==23.2.1

# Data validation & settings
pydantic==2.5.0
pydantic-settings==2.1.0

# Google APIs
google-api-python-client==2.116.0
google-auth==2.27.0

# HTTP client (async)
httpx==0.26.0

# CLI
click==8.1.7

# Logging
structlog==24.1.0

# Image handling
pillow==10.2.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
respx==0.20.2
